feat: flow without mbs

+++ MyFinance-Backend/MyFinance.Application/UseCases/Transfers/Commands/RegisterTransfer/RegisterTransferCommand.cs
@@ -8,7 +8,7 @@ namespace MyFinance.Application.UseCases.Transfers.Commands.RegisterTransfer;
 public sealed record RegisterTransferCommand(
     Guid BusinessUnitId,
     Guid AccountTagId,
-    double Value,
+    decimal Value,
     string RelatedTo,
     string Description,
     DateTime SettlementDate,
