+++ MyFinance-Backend/MyFinance.Contracts/Transfer/Requests/RegisterTransferRequest.cs
@@ -6,7 +6,7 @@ public sealed class RegisterTransferRequest
 {
     public required Guid BusinessUnitId { get; init; }
     public required Guid AccountTagId { get; init; }
-    public required double Value { get; init; }
+    public required decimal Value { get; init; }
     public required string RelatedTo { get; init; }
     public required string Description { get; init; }
     public required DateTime SettlementDate { get; init; }
