+++ MyFinance-Backend/MyFinance.Contracts/Transfer/Requests/UpdateTransferRequest.cs
@@ -6,7 +6,7 @@ public sealed class UpdateTransferRequest
 {
     public required Guid Id { get; init; }
     public required Guid AccountTagId { get; init; }
-    public required double Value { get; init; }
+    public required decimal Value { get; init; }
     public required string RelatedTo { get; init; }
     public required string Description { get; init; }
     public required DateTime SettlementDate { get; init; }
